server: renamed trpc router files to <feature>.trpc.ts
