+++ apps/portal/src/integrations/trpc.ts
@@ -1,5 +1,5 @@
 import { createTRPCContext } from '@trpc/tanstack-react-query'
 
-import type { AppRouter } from '@server/app/trpc/trpc.app-router'
+import type { AppRouter } from '@server/app/trpc/app.trpc'
 
 export const { TRPCProvider, useTRPC } = createTRPCContext<AppRouter>()
