+++ apps/server/src/app/http/app.ts
@@ -2,8 +2,8 @@ import * as trpcExpress from '@trpc/server/adapters/express'
 import express from 'express'
 import { logger } from 'firebase-functions/v2'
 
+import { appRouter } from '@/app/trpc/app.trpc'
 import { createContext } from '@/app/trpc/trpc'
-import { appRouter } from '@/app/trpc/trpc.app-router'
 import { getErrorCode, type AppErrorCode } from '@/app/trpc/trpc.errors'
 import { googleReviewsRoute } from '@/features/google-business-profile/functions/routes/google-reviews'
 import { invitationEntryRedirect } from '@/features/party-bookings/functions/webhooks/invitation-redirect'
