+++ apps/server/src/features/auth/functions/trpc/auth.trpc.ts
