+++ apps/server/src/features/events/functions/trpc/events.trpc.ts
