+++ apps/server/src/features/party-bookings/functions/trpc/parties.trpc.ts
